requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database Connectivity
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy-redshift==0.8.14
boto3==1.34.0
pandas==2.1.3
pyarrow==14.0.1
numpy<2.0
google-cloud-bigquery==3.13.0
sqlalchemy-bigquery==1.9.0

# AI and LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
tiktoken==0.5.2
watchfiles==0.21.0

# Embeddings and Vector Stores
sentence-transformers==2.2.2
faiss-cpu==1.8.0
chromadb==0.4.18

# UI and Visualization
streamlit==1.29.0
plotly==5.18.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Optional - Local LLM Support
# ollama==0.1.5

# Optional - Production Vector Stores
# pinecone-client==2.2.4
# weaviate-client==3.25.3
